Module PostgresORM

PostgresORM.retrieve_one_entity — Function

retrieve_one_entity(filter_object::T,
                    retrieve_complex_props::Bool,
                    dbconn::LibPQ.Connection)

Retrieves an instance ot T that matches the filter. If there is no match then the function returns missing. If there are more than one match, an error is thrown. If retrieve_complex_props is true then a complex prop (i.e. a property of type IEntity) is fully loaded using an additional query to the database ; if not, an instance of IEntity is simply created using the values of the foreign key

PostgresORM.retrieve_entity — Function

retrieve_entity(filter_object::IEntity,
                retrieve_complex_props::Bool,
                dbconn::LibPQ.Connection)

Retrieves a vector or T instances that match the filter. If there is no match, it returns an empty vector.